regressor refactor
