+++ aeon/clustering/deep_learning/__init__.py
@@ -1,4 +1,4 @@
 """Deep learning based clusterers."""
 __all__ = ["BaseDeepClusterer", "AEFCNClusterer"]
-from aeon.clustering.deep_learning.ae_fcn import AEFCNClusterer
+from aeon.clustering.deep_learning._ae_fcn import AEFCNClusterer
 from aeon.clustering.deep_learning.base import BaseDeepClusterer
