+++ aeon/regression/deep_learning/_fcn.py
@@ -10,9 +10,9 @@ from copy import deepcopy
 
 from sklearn.utils import check_random_state
 
-from aeon.networks._fcn import FCNNetwork
+from aeon.networks import FCNNetwork
 from aeon.regression.deep_learning.base import BaseDeepRegressor
-from aeon.utils.validation._dependencies import _check_dl_dependencies
+from aeon.utils.validation._dependencies import _check_soft_dependencies
 
 
 class FCNRegressor(BaseDeepRegressor):
@@ -130,7 +130,7 @@ class FCNRegressor(BaseDeepRegressor):
         use_bias=True,
         optimizer=None,
     ):
-        _check_dl_dependencies(severity="error")
+        _check_soft_dependencies("tensorflow")
         super(FCNRegressor, self).__init__(last_file_name=last_file_name)
 
         self.n_layers = n_layers
