+++ aeon/regression/deep_learning/_tapnet.py
@@ -12,9 +12,9 @@ from copy import deepcopy
 
 from sklearn.utils import check_random_state
 
-from aeon.networks._tapnet import TapNetNetwork
+from aeon.networks import TapNetNetwork
 from aeon.regression.deep_learning.base import BaseDeepRegressor
-from aeon.utils.validation._dependencies import _check_dl_dependencies
+from aeon.utils.validation._dependencies import _check_soft_dependencies
 
 
 class TapNetRegressor(BaseDeepRegressor):
@@ -106,7 +106,7 @@ class TapNetRegressor(BaseDeepRegressor):
         callbacks=None,
         verbose=False,
     ):
-        _check_dl_dependencies(severity="error")
+        _check_soft_dependencies("tensorflow")
         super(TapNetRegressor, self).__init__()
 
         self.batch_size = batch_size
